Guess the session
Sailfish OS uses systemd userspace daemon.
The problem about systemd userspace daemon
is that any process started by systemd
will not be attached to a session. This
means that the polkit agent will not have
a session either.

This commit will try to find the first
(only) session belonging to current user
(aka nemo) that also have a seat, that
is the only session that have been
opened on the device.